파일 업로드 수정

+++ savg/src/main/java/SeAH/savg/service/EduService.java
@@ -159,18 +159,9 @@ public class EduService {
         Edu edu = eduRepository.findByEduId(eduId);
 
         List<EduFile> eduFileList = eduFileRepository.findByEdu(edu);
-        List<EduFileDTO> eduFileDTOList = new ArrayList<>();
         EduDTO eduDTO = new EduDTO(edu);
-        List<String> filesName = new ArrayList<>();
-        for(EduFile file : eduFileList){
-            EduFileDTO eduFileDTO = EduFileDTO.of(file);
-            eduFileDTOList.add(eduFileDTO);
-         filesName.add(eduFileDTO.getEduFileOriName());
 
-        }
-        eduDTO.setEduFiles(filesName);
-
-
+        eduDTO.setEduFileList(eduFileList);
 
         return eduDTO;
     }
@@ -179,11 +170,13 @@ public class EduService {
         Edu edu = eduDTO.toEntity();
         List<EduFile> eduFileList = eduFileRepository.findByEdu(edu);
 
-        for(EduFile eduFile : eduFileList){
-            eduFileRepository.delete(eduFile);
+        if(eduDTO.getFiles()!=null){
+            for(EduFile eduFile : eduFileList){
+                eduFileRepository.delete(eduFile);
+            }
+            eduFileService.uploadFile(eduDTO);
+            eduDTO.setEduFileList(eduFileList);
         }
-
-        eduFileService.uploadFile(eduDTO);
 
         eduRepository.save(edu);
     }
